@@ -1,25 +1,25 @@
 {
-  "name": "mongddang-front",
+  "name": "frontend",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@tanstack/react-query": "^4.23.0",
-    "@tanstack/react-query-devtools": "^4.23.0",
+    "@tanstack/react-query": "^4.26.1",
+    "@tanstack/react-query-devtools": "^4.26.1",
     "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^12.1.4",
+    "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
-    "axios": "^1.2.3",
-    "react": "^17.0.2",
+    "axios": "^1.3.4",
+    "react": "^18.2.0",
     "react-cookie": "^4.1.1",
-    "react-dom": "^17.0.2",
-    "react-intersection-observer": "^9.4.2",
-    "react-query": "^3.39.3",
-    "react-router-dom": "^6.6.2",
+    "react-dom": "^18.2.0",
+    "react-intersection-observer": "^9.4.3",
+    "react-router-dom": "^6.9.0",
     "react-scripts": "5.0.1",
-    "styled-components": "^5.3.6",
+    "styled-components": "^5.3.9",
     "twin.macro": "^3.1.0",
     "web-vitals": "^2.1.4",
-    "zustand": "^4.3.2"
+    "yarn": "^1.22.19",
+    "zustand": "^4.3.6"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -48,9 +48,9 @@
   "devDependencies": {
     "babel-plugin-macros": "^3.1.0",
     "babel-plugin-styled-components": "^2.0.7",
-    "eslint-config-prettier": "^8.6.0",
+    "eslint-config-prettier": "^8.7.0",
     "eslint-plugin-prettier": "^4.2.1",
-    "prettier": "^2.8.2",
-    "tailwindcss": "^3.2.4"
+    "prettier": "^2.8.4",
+    "tailwindcss": "^3.2.7"
   }
 }
